+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.4 - 2023-02-22
+
+### Added
+- Custom overlay's line width, font size, and color.
+- A conditional color based on measured angle
+
+
 ## v0.3 - 2023-02-14
 
 ### Added